+++ _components/bypasslinks/bypasslinks.js
@@ -1,9 +1,5 @@
 'use strict';
 
-// Set Array prototype on NodeList for forEach() support
-// https://gist.github.com/paulirish/12fb951a8b893a454b32#gistcomment-1474959
-NodeList.prototype.forEach = Array.prototype.forEach;
-
 /**
  * @param {object} options Object containing configuration overrides
  */
@@ -14,6 +10,7 @@ const Frbypasslinks = function({
 
 	//	CONSTANTS
 	const doc = document;
+	const _q = (el, ctx = doc) => [].slice.call(ctx.querySelectorAll(el));
 
 
 	//	SUPPORTS
@@ -22,7 +19,7 @@ const Frbypasslinks = function({
 
 	//	SETUP
 	// get bypass links NodeList
-	const container = doc.querySelector(selector);
+	const container = _q(selector)[0];
 
 	//	TEMP
 	let currTarget = null;
@@ -34,20 +31,18 @@ const Frbypasslinks = function({
 		let id = link.getAttribute('href').slice(1);
 		let target = doc.getElementById(id);
 		//	set tabindex to allow focus
-		target.setAttribute('tabindex', -1);
+		if (target) target.setAttribute('tabindex', -1);
 	}
 	function _removeFocusability (link) {
 		//	get target element
 		let id = link.getAttribute('href').slice(1);
 		let target = doc.getElementById(id);
 		//	remove ability to focus (stops user highlighting element on click)
-		target.removeAttribute('tabindex');
+		if (target) target.removeAttribute('tabindex');
 	}
 	function destroy () {
-		//	get all bypass links
-		const links = container.querySelectorAll('a');
 		//	loop through each bypass link and remove event bindings
-		links.forEach(link => {
+		_q('a', container).forEach(link => {
 			_unbindPointerClick(link)
 			_addFocusability(link);
 		});
@@ -60,6 +55,9 @@ const Frbypasslinks = function({
 		//	get target element
 		let id = e.target.getAttribute('href').slice(1);
 		let target = doc.getElementById(id);
+
+		// don't try to apply relevant atts/focus if target isn't present
+		if (!target) return;
 		//	set tabindex to allow focus
 		target.setAttribute('tabindex', -1);
 		target.focus();
@@ -104,10 +102,8 @@ const Frbypasslinks = function({
 	function init () {
 		//	detect if bypass links exist in the document
 		if (!container) return;
-		//	get all bypass links
-		const links = container.querySelectorAll('a');
 		//	loop through each bypass link
-		links.forEach(link => {
+		_q('a', container).forEach(link => {
 			_bindPointerClick(link);
 			_removeFocusability(link);
 		});