+++ _components/offcanvas/offcanvas.js
@@ -1,9 +1,5 @@
 'use strict';
 
-// Set Array prototype on NodeList for forEach() support
-// https://gist.github.com/paulirish/12fb951a8b893a454b32#gistcomment-1474959
-NodeList.prototype.forEach = Array.prototype.forEach;
-
 // Polyfill matches as per https://github.com/jonathantneal/closest
 Element.prototype.matches = Element.prototype.matches ||
 							Element.prototype.mozMatchesSelector ||
@@ -26,6 +22,7 @@ const Froffcanvas = function({
 	//	CONSTANTS
 	const doc = document;
 	const docEl = doc.documentElement;
+	const _q = (el, ctx = doc) => [].slice.call(ctx.querySelectorAll(el));
 
 
 	//	SUPPORTS
@@ -34,7 +31,7 @@ const Froffcanvas = function({
 
 	//	SETUP
 	// set offcanvas element NodeLists
-	const panels = doc.querySelectorAll(selector);
+	const panels = _q(selector);
 
 	//	TEMP
 	let currButtonOpen = null;
@@ -129,12 +126,12 @@ const Froffcanvas = function({
 	//	EVENTS
 	function _eventOpenPointer (e) {
 		//	get panel
-		let panelId = e.target.getAttribute('aria-controls');
-		let panel = doc.querySelector(`#${panelId}`);
+		let panelId = e.currentTarget.getAttribute('aria-controls');
+		let panel = doc.getElementById(panelId);
 		//	hide any open panels
 		if (currPanel) _hidePanel(currPanel, false);
 		//	save temp panel/button
-		currButtonOpen = e.target;
+		currButtonOpen = e.currentTarget;
 		currPanel = panel;
 		//	show
 		_showPanel(panel);
@@ -156,11 +153,11 @@ const Froffcanvas = function({
 
 	//	BIND EVENTS
 	function _bindOpenPointer (panel) {
-		const openButtons = doc.querySelectorAll(`${openSelector}[aria-controls="${_getPanelId(panel)}"]`); // is this selector totally crazy?
+		const openButtons = _q(`${openSelector}[aria-controls="${_getPanelId(panel)}"]`); // is this selector totally crazy?
 		openButtons.forEach((button) => button.addEventListener('click', _eventOpenPointer));
 	}
 	function _bindClosePointer (panel = currPanel) {
-		var closeButton = panel.querySelector(closeSelector);
+		var closeButton = _q(closeSelector, panel)[0];
 		closeButton.addEventListener('click', _eventClosePointer);
 	}
 	function _bindDocClick () {
@@ -173,11 +170,11 @@ const Froffcanvas = function({
 
 	//	UNBIND EVENTS
 	function _unbindOpenPointer (panel = currPanel) {
-		const openButtons = doc.querySelectorAll(`${openSelector}[aria-controls="${_getPanelId(panel)}"]`); // yep its totally crazy
+		const openButtons = _q(`${openSelector}[aria-controls="${_getPanelId(panel)}"]`); // yep its totally crazy
 		openButtons.forEach((button) => button.removeEventListener('click', _eventOpenPointer));
 	}
 	function _unbindClosePointer (panel = currPanel) {
-		var closeButton = panel.querySelector(closeSelector);
+		var closeButton = _q(closeSelector, panel)[0];
 		closeButton.removeEventListener('click', _eventClosePointer);
 	}
 	function _unbindDocClick () {