+++ _components/tabs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Frend Tabs Change Log
 
+## [1.0.5] 2016-06-19
+- Improve DOM traversal with `_closest()` utility
+- Add `matches` polyfill for better browser support
+
+## [1.0.4] 2016-06-05
+- Replace NodeList.prototype jacking with `_q()` querySelectorAll util
+
+## [1.0.3] 2016-06-03
+- Added guard clause for modifier key events - fix issue 75
+
 ## [1.0.2] 2016-05-22
 - Added aria-labelledby for tabpanels
 - Improve ids on tab anchors and tabpanels